🔨 refactor: type, reducer 파일 분리
