+++ src/hooks/useClipboard/type.ts
@@ -1,0 +1,9 @@
+export interface UseClipboardProps {
+  resetTime?: number;
+}
+
+export interface UseClipboardReturns {
+  copied: boolean;
+  copyText: (text: string) => void;
+  copyImg: (path: string) => void;
+}
