+++ src/hooks/useDebounce/type.ts
@@ -1,0 +1,1 @@
+export type GenericFn<T extends unknown[]> = (...args: T) => void;
