+++ src/hooks/useDelayFlag/type.ts
@@ -1,0 +1,6 @@
+export type PositiveInteger<T extends number> = `${T}` extends
+  | `-${string}`
+  | `${string}.${string}`
+  | '0'
+  ? never
+  : T;
