+++ src/hooks/useDetectInactivity/type.ts
@@ -1,0 +1,1 @@
+export type Fn = () => void;
