+++ src/hooks/useGeolocation/type.ts
@@ -1,0 +1,6 @@
+export interface UseGeolocationReturnType
+  extends Partial<GeolocationCoordinates> {
+  loading: boolean;
+  error: GeolocationPositionError | null;
+  timestamp: EpochTimeStamp | undefined;
+}
