+++ src/hooks/useHover/type.ts
@@ -1,0 +1,4 @@
+export interface UseHoverReturns<T> {
+  isHovered: boolean;
+  callbackRef: (node: T) => void;
+}
