+++ src/hooks/useKeyCombination/type.ts
@@ -1,0 +1,7 @@
+export type Fn = () => void;
+
+export interface UseKeyCombinationProps {
+  shortcutKeys: string[];
+  callback: Fn;
+  isPrevent?: boolean;
+}
