+++ src/hooks/useLocalStorage/type.ts
@@ -1,0 +1,1 @@
+export type ValueResolver<T> = T | ((prevPayload: T) => T);
