+++ src/hooks/useOnlineStatus/type.ts
@@ -1,0 +1,10 @@
+export type Fn = () => void;
+
+export interface useOnlineStatusProps {
+  onlineCallback?: Fn;
+  offlineCallback?: Fn;
+}
+
+export interface UseOnlineStatusReturns {
+  isOnline: boolean;
+}
