+++ src/hooks/useOutsideInteraction/type.ts
@@ -1,0 +1,11 @@
+export type EventType =
+  | 'mousedown'
+  | 'mouseup'
+  | 'touchstart'
+  | 'touchend'
+  | 'keydown';
+
+export interface UseOutsideInteractionProps {
+  handleOutsideInteraction: () => void;
+  events?: EventType[];
+}
