+++ src/hooks/usePrefersColorScheme/type.ts
@@ -1,0 +1,7 @@
+export type Fn = () => void;
+
+export type ColorSchemeType = 'dark' | 'light';
+
+export interface UsePrefersColorSchemeProps {
+  serverSnapshot?: ColorSchemeType;
+}
