+++ src/hooks/useTranslation/type.ts
@@ -1,0 +1,5 @@
+export interface UseTranslationReturns {
+  t: (key: string, variables?: { [key: string]: string | number }) => string;
+  language: string;
+  changeLanguage: (lang: string) => void;
+}
