+++ src/hooks/useUnmountEffect/useUnmountEffect.ts
@@ -1,12 +1,11 @@
 import { useEffect, useRef } from 'react';
+import { Fn } from './type';
 
 /**
  * 컴포넌트가 언마운트 될 때 전달받은 인자 함수를 호출하는 훅
  *
  * @param callback 언마운트 시에 호출될 함수
  */
-
-type Fn = () => void;
 
 const useUnmountEffect = (callback: Fn) => {
   const callbackRef = useRef<null | Fn>(callback);
