+++ src/hooks/useWindowSize/type.ts
@@ -1,0 +1,11 @@
+export type PositiveInteger<T extends number> = `${T}` extends
+  | `-${string}`
+  | `${string}.${string}`
+  | '0'
+  ? never
+  : T;
+
+export interface UseWindowSizeReturns {
+  width: number | null;
+  height: number | null;
+}
