+++ src/hooks/useWorker/type.ts
@@ -1,0 +1,1 @@
+export type WorkerScript<A, R, C> = (arg: A, accessClosure?: C) => R;
